+++ pkgs/cupertino_http/test/main.dart
@@ -6,6 +6,7 @@ import 'client_conformance_test.dart' as client_conformance_test;
 import 'client_profile_test.dart' as profile_test;
 import 'client_test.dart' as client_test;
 import 'http_url_response_test.dart' as http_url_response_test;
+import 'isolate_shutdown_test.dart' as isolate_shutdown_test;
 import 'mutable_url_request_test.dart' as mutable_url_request_test;
 import 'url_cache_test.dart' as url_cache_test;
 import 'url_request_test.dart' as url_request_test;
@@ -23,10 +24,11 @@ import 'web_socket_conformance_test.dart' as web_socket_conformance_test;
 /// application launch is required.
 void main() {
   client_conformance_test.main();
-  profile_test.main();
   client_test.main();
   http_url_response_test.main();
+  isolate_shutdown_test.main();
   mutable_url_request_test.main();
+  profile_test.main();
   url_cache_test.main();
   url_request_test.main();
   url_response_test.main();