+++ pkgs/cupertino_http/CHANGELOG.md
@@ -1,6 +1,13 @@
-## 3.0.3-wip
+## 3.1.1-wip
 
 * Update example to fetch and display scores of package:cupertino_http from pub.dev.
+* Update to `package:ffigen` `22.0.0`.
+
+## 3.1.0
+
+* Switch to `package:objective_c` `9.5.0` and `package:ffigen` `21.0.0`, fixing
+  a [bug](https://github.com/dart-lang/http/issues/1894) where a request that
+  was still in flight when its isolate shut down could crash the process.
 
 ## 3.0.2